feat: split code in different files
